+++ app/models/user.rb
@@ -13,7 +13,7 @@ class User
   validates_format_of :email, as: :email_address
 
   property :id, Serial
-  property :email, String, required: true
+  property :email, String, required: true, unique: true
   property :password_digest, String, length: 60
