+++ app/views/layout.erb
@@ -5,9 +5,18 @@
   </head>
 
   <body>
-    <% if flash[:notice] %>
+    <!-- <% if flash[:notice] %>
       <div id='notice'><%= flash[:notice] %></div>
+    <% end %> -->
+
+    <% if flash[:errors] && !flash[:errors].empty? %>
+  Please refer to the following errors below:
+  <ul id='errors'>
+    <% flash[:errors].each do |error| %>
+      <li><%= error %></li>
     <% end %>
+  </ul>
+<% end %>
 
     <% if current_user %>
     Welcome, <%= current_user.email %>
