+++ spec/features/sign_in_spec.rb
@@ -9,11 +9,4 @@ feature 'User sign in' do
     expect(page.status_code).to eq(200)
     expect(page).to have_content("Welcome, #{user.email}")
   end
-
-  def sign_in(email:, password:)
-    visit('/sessions/new')
-    fill_in :email, with: email
-    fill_in :password, with: password
-    click_button('Log in')
-  end
 end
