+++ spec/features/sign_up_spec.rb
@@ -12,23 +12,24 @@ feature 'User Sign Up' do
   scenario 'with a password that does not match' do
     expect { sign_up(password_confirmation: 'wrong') }.not_to change(User, :count)
     expect(current_path).to eq('/new')
-    expect(page).to have_content('password and confirmation do not match')
+    expect(page).to have_content('Password does not match the confirmation')
   end
 
   scenario 'you cannot sign up without an email address' do
     expect { sign_up(email: nil) }.not_to change(User, :count)
+    expect(current_path).to eq('/new')
+    expect(page).to have_content('Email must not be blank')
   end
 
   scenario 'you cannot sign up without a valid email address' do
     expect { sign_up(email: 'invalid@email')}.not_to change(User, :count)
+    expect(current_path).to eq('/new')
+    expect(page).to have_content('Email has an invalid format')
   end
 
-def sign_up(email: 'tim.chipperfield@gmail.com', password: 'christmastrees', password_confirmation: 'christmastrees' )
-    visit '/'
-    expect(page.status_code).to eq(200)
-    fill_in :email, with: email
-    fill_in :password, with: password
-    fill_in :password_confirmation, with: password_confirmation
-    click_button 'Sign Up'
+  scenario "I cannot sign up with an existing email address" do
+    sign_up
+    expect { sign_up }.to_not change(User, :count)
+    expect(page).to have_content('Email is already taken')
   end
 end
